HZN-1063: Fixing some tests
